refactor btcd
Former-commit-id: b577d69d1903b49848128868fdee9c9f0d375dd8
Former-commit-id: 56b68a37e94a9fe0ed1be48c1f346cab213e5c77

+++ cmd/relayd/relayd/btcweb.go
@@ -114,7 +114,7 @@ func (b *BtcWeb) GetSPV(height uint64, txHash string) (*types.BtcSpv, error) {
 		Time:        block.Time,
 		Height:      block.Height,
 		BlockHash:   block.Hash,
-		TxIndex:     uint64(txIndex),
+		TxIndex:     txIndex,
 		BranchProof: proof,
 	}
 	return spv, nil
